+++ src/vegas_mc/VegasMC.jl
@@ -6,8 +6,6 @@ import ..Configuration
 import ..MPIreduceConfig!
 import ..addConfig!
 import ..clearStatistics!
-import ..setWeight!
-import ..initialize!
 
 import ..TINY
